requirements.txt
openai>=1.0.0
httpx>=0.27.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.21.0
pandas>=1.3.0
PyPDF2>=3.0.0
Pillow>=8.0.0
PyMuPDF>=1.23.0
streamlit>=1.20.0
streamlit-ace>=0.1.0
pdfplumber>=0.10.0
pdf2image>=1.16.0
requests>=2.28.0